pyproject.toml
[project]
name = "gv_point_cloud_exporter"
version = "0.0.4"
description = "A Blender addon for exporting PointCloud and Splat data to PLY format."
authors = ["Giancarlo Viali"]
license = "MIT"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End-Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
]
dependencies = [
    "fake-bpy-module>=20231205",
]